Resolve "Correct Int Literal Cast Warnings From Error Prone"

Closes #1577

See merge request echothree/echothree!1241

+++ src/java/com/echothree/control/user/chain/server/command/ProcessChainInstanceStatusesCommand.java
@@ -51,7 +51,7 @@ public class ProcessChainInstanceStatusesCommand
         var chainInstanceStatusLogic = ChainInstanceStatusLogic.getInstance();
         BasePK processedBy = getPartyPK();
         long chainInstanceStatusesProcessed = 0;
-        var remainingTime = (long) 2 * 60 * 1000; // 2 minutes
+        var remainingTime = 2L * 60 * 1000; // 2 minutes
         
         for(var chainInstanceStatus: chainControl.getChainInstanceStatusesByNextChainActionSetTimeForUpdate(session.START_TIME_LONG)) {
             var startTime = System.currentTimeMillis();
